ai-server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.51.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic>=2.9.0
